+++ MyLibraries/Ili9320/ili9320.h
@@ -31,8 +31,8 @@
  */
 
 void ILI9320_Initializtion(void);
-void ILI9320_SetWindow(uint16_t x, uint16_t y, uint16_t width, uint16_t height);
-void ILI9320_DrawPixel(uint16_t x, uint16_t y, uint8_t r, uint8_t g, uint8_t b);
+void ILI9320_SetWindow(int x, int y, int width, int height);
+void ILI9320_DrawPixel(int x, int y, uint8_t r, uint8_t g, uint8_t b);
 uint16_t ILI9320_RGBDecode(uint8_t r, uint8_t g, uint8_t b);
 
 /**
